update allocation refactor
