+++ src/components/EditAllocationRate.js
@@ -27,106 +27,46 @@ const EditAllocationRate = (props) => {
         allocation,
         currency,
         rate,
-        onClose
+        //onClose,
+
+        endDate,
+        setEndDate,
+        setNewAllocationRate,
+        setSelectedCurrency,
+        setStartDate,
+        startDate
     } = props
 
-    const {
-        data: dataContributorRates,
-        loading: loadingContributorRates,
-        error: errorContributorRates
-    } = useQuery(GET_CONTRIBUTOR_RATES, {
-        variables: {
-            id: allocation.contributor.id
-        }
-    })
+    // const [createRate, {
+    //     dataNewRate,
+    //     loadingNewRate,
+    //     errorNewRate
+    // }] = useMutation(CREATE_RATE)
 
-    const [createRate, {
-        dataNewRate,
-        loadingNewRate,
-        errorNewRate
-    }] = useMutation(CREATE_RATE)
+    // const [updateAllocation, {
+    //     dataUpdatedAllocation,
+    //     loadingUpdatedAllocation,
+    //     errorUpdatedAllocation
+    // }] = useMutation(UPDATE_ALLOCATION, {
+    //     refetchQueries: [{
+    //         query: GET_CONTRIBUTOR_ALLOCATIONS,
+    //         variables: {
+    //             id: allocation.contributor.id
+    //         }
+    //     }]
+    // })
 
-    const [updateAllocation, {
-        dataUpdatedAllocation,
-        loadingUpdatedAllocation,
-        errorUpdatedAllocation
-    }] = useMutation(UPDATE_ALLOCATION, {
-        refetchQueries: [{
-            query: GET_CONTRIBUTOR_ALLOCATIONS,
-            variables: {
-                id: allocation.contributor.id
-            }
-        }]
-    })
-
-    const [endDate, setEndDate] = useState(moment(allocation.end_date, 'x')['_d'])
-    const [newAllocationRate, setNewAllocationRate] = useState({})
-    const [selectedCurrency, setSelectedCurrency] = useState(null)
+    //const [endDate, setEndDate] = useState(moment(allocation.end_date, 'x')['_d'])
+    //const [newAllocationRate, setNewAllocationRate] = useState({})
+    //const [selectedCurrency, setSelectedCurrency] = useState(null)
     const [selectedRateType, setSelectedRateType] = useState(rate.type)
-    const [startDate, setStartDate] = useState(moment(allocation.start_date, 'x')['_d'])
+    //const [startDate, setStartDate] = useState(moment(allocation.start_date, 'x')['_d'])
 
     const getRangedTimeEntries = (dates) => {
         const [start, end] = dates
         setStartDate(start)
         setEndDate(end)
     }
-    const handleUpdateAllocation = async ({
-        allocation,
-        contributor,
-        contributorRates,
-        rate,
-        startDate,
-        endDate
-    }) => {
-        //look for rate with same values
-        const selectedRate = {}
-        const existingRate = findKey(
-            contributorRates,
-            {
-                'hourly_rate': rate.hourly_rate.toString(),
-                'total_expected_hours': Number(rate.total_expected_hours),
-                'type': rate.type,
-                'currency': selectedCurrency
-            }
-        )
-        if (existingRate) {
-            selectedRate.id = contributorRates[existingRate].id
-        } else {
-            //create rate
-            const newRate = await createRate({
-                variables: {
-                    hourly_rate: rate.hourly_rate.toString(),
-                    total_expected_hours: Number(rate.total_expected_hours),
-                    type: rate.type,
-                    currency: selectedCurrency,
-                    contributor_id: contributor.id
-                }
-            })
-            selectedRate.id = newRate.data.createRate.id
-        }
-        //update allocation with that rate id
-        const updatedAllocation = await updateAllocation({
-            variables: {
-                id: allocation.id,
-                amount: Number(rate.total_amount),
-                start_date: moment(startDate).format('YYYY-MM-DD'),
-                end_date: moment(endDate).format('YYYY-MM-DD'),
-                date_paid: null,
-                rate_id: Number(selectedRate.id)
-            }
-        })
-        if (loadingUpdatedAllocation) return ''
-        else if (updatedAllocation.errors) {
-            console.log('Error updating the allocation');
-        } else {
-            onClose()
-        }
-    }
-
-    if (loadingContributorRates) return <LoadingProgress/>
-    if (errorContributorRates) return `${errorContributorRates}`
-
-    const { getContributorById: contributorRates } = dataContributorRates
 
     return (
         <Box>
@@ -197,30 +137,13 @@ const EditAllocationRate = (props) => {
                             currentTotal={allocation.amount}
                             clientCurrency={currency}
                             currentRate={rate}
+                            setCurrency={setSelectedCurrency}
                             setNewAllocationRate={setNewAllocationRate}
                             startDate={moment(startDate)}
                             endDate={moment(endDate)}
                         />
                     )
             }
-            <Grid container>
-                <Grid item xs={3}>
-                    <Button
-                        variant='contained'
-                        color='primary'
-                        onClick={() => handleUpdateAllocation({
-                            allocation: allocation,
-                            contributor: allocation.contributor,
-                            contributorRates: contributorRates.rates,
-                            rate: newAllocationRate,
-                            startDate: startDate,
-                            endDate: endDate
-                        })}
-                    >
-                        {'Edit'}
-                    </Button>
-                </Grid>
-            </Grid>
         </Box>
     )
